+++ sort.h
@@ -48,15 +48,15 @@ extern int_fast32_t sort_next(sort_t *sort);
 extern void         sort_init_qs(sort_t *sort, board_t *board, const attack_t *attack, bool check);
 extern int_fast32_t sort_next_qs(sort_t *sort);
 
-extern void good_move(int_fast32_t move, const board_t *board, int_fast32_t depth, int_fast32_t height);
+extern void good_move(uint_fast16_t move, const board_t *board, int_fast32_t depth, int_fast32_t height);
 
-extern void history_good(int_fast32_t move, const board_t *board);
-extern void history_bad(int_fast32_t move, const board_t *board);
-extern void history_very_bad(int_fast32_t move, const board_t *board);
+extern void history_good(uint_fast16_t move, const board_t *board);
+extern void history_bad(uint_fast16_t move, const board_t *board);
+extern void history_very_bad(uint_fast16_t move, const board_t *board);
 
-extern bool history_reduction(int_fast32_t move, const board_t *board);
-extern void history_tried(int_fast32_t move, const board_t *board);
-extern void history_success(int_fast32_t move, const board_t *board);
+extern bool history_reduction(uint_fast16_t move, const board_t *board);
+extern void history_tried(uint_fast16_t move, const board_t *board);
+extern void history_success(uint_fast16_t move, const board_t *board);
 
 extern void note_moves(list_t *list, const board_t *board, int_fast32_t height, int_fast32_t trans_killer);
